Add session resuming

+++ LavalinkServer/src/main/java/lavalink/server/io/WebSocketHandlers.kt
@@ -105,7 +105,7 @@ class WebSocketHandlers(socketServer: SocketServer) {
     fun configureResuming(session: WebSocketSession, json: JSONObject) {
         val socketContext = contextMap[session.id]!!
         socketContext.resumeKey = json.optString("key")
-        if (json.has("timeout")) socketContext.resumeTimeout = json.getInt("timeout")
+        if (json.has("timeout")) socketContext.resumeTimeout = json.getLong("timeout")
     }
 
 }
